refactor(Establishment, Services): Doing the pages server side to handle the admin role better
