+++ frontend/src/app/forbidden/page.jsx
@@ -1,0 +1,11 @@
+"use server";
+import Layout from "../components/base_layout/Layout";
+
+export default async function forbidden(){
+
+  return (
+    <Layout>
+        <title>SG BarberShop</title>
+    </Layout>
+  );
+}
