+++ frontend/src/app/utils/requestBuilder.js
@@ -28,6 +28,21 @@ export const getId = async () => {
     }
 };
 
+export const isAdmin = async () => {
+    try{
+        
+        const cookieStore = await cookies();
+        const token = cookieStore.get("token")?.value;
+        const payload = JSON.parse(
+            Buffer.from(token.split('.')[1], 'base64').toString()
+        );
+
+        return payload.role == "admin";
+    } catch(err){
+        return false;
+    }
+}
+
 export const logOut = async () => {
     try{
         const cookieStore = await cookies();
